refactor: modularise additional functions
